// src/api/schema.d.ts
export interface components {
    schemas: {
        AdminCreate: {
            readonly id: number;
            first_name: string;
            last_name: string;
            readonly role: components["schemas"]["Role"];
            role_id: number;
            phone: string;
            password: string;
            /** Format: date-time */
            readonly updated_at: string;
            /** Format: date-time */
            readonly created_at: string;
            /**
             * Oxirgi kirish
             * Format: date-time
             */
            last_login?: string | null;
        };
        AdminGetMe: {
            readonly id: number;
            first_name: string;
            last_name: string;
            readonly full_name: string;
            readonly role: components["schemas"]["Role"];
            readonly phone: string;
            /**
             * Oxirgi kirish
             * Format: date-time
             */
            readonly last_login: string | null;
            readonly is_staff: boolean;
            readonly is_superuser: boolean;
            /** Format: date-time */
            readonly created_at: string;
        };
// ...
        PaginatedAdminCreateList: {
            data: components["schemas"]["AdminCreate"][];
            pagination: {
                /** @example 123 */
                count: number;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=4
                 */
                next?: string | null;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=2
                 */
                previous?: string | null;
            };
        };
// ...
        PaginatedRoleList: {
            data: components["schemas"]["Role"][];
            pagination: {
                /** @example 123 */
                count: number;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=4
                 */
                next?: string | null;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=2
                 */
                previous?: string | null;
            };
        };
        PaginatedUserSalaryList: {
            data: components["schemas"]["UserSalary"][];
            pagination: {
                /** @example 123 */
                count: number;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=4
                 */
                next?: string | null;
                /**
                 * Format: uri
                 * @example http://api.example.org/accounts/?page=2
                 */
                previous?: string | null;
            };
        };
// ...
        PatchedAdminCreate: {
            readonly id?: number;
            first_name?: string;
            last_name?: string;
            readonly role?: components["schemas"]["Role"];
            role_id?: number;
            phone?: string;
            password?: string;
            /** Format: date-time */
            readonly updated_at?: string;
            /** Format: date-time */
            readonly created_at?: string;
            /**
             * Oxirgi kirish
             * Format: date-time
             */
            last_login?: string | null;
        };
        PatchedAdminGetMe: {
            readonly id?: number;
            first_name?: string;
            last_name?: string;
            readonly full_name?: string;
            readonly role?: components["schemas"]["Role"];
            readonly phone?: string;
            /**
             * Oxirgi kirish
             * Format: date-time
             */
            readonly last_login?: string | null;
            readonly is_staff?: boolean;
            readonly is_superuser?: boolean;
            /** Format: date-time */
            readonly created_at?: string;
        };
// ...
        PatchedRole: {
            readonly id?: number;
            code?: string;
            name?: string;
        };
        PatchedUserGetMe: {
            readonly id?: number;
            first_name?: string;
            last_name?: string;
            readonly full_name?: string;
            readonly role?: components["schemas"]["Role"];
            readonly phone?: string;
            /** Format: date-time */
            readonly created_at?: string;
        };
        PatchedUserSalary: {
            readonly id?: number;
            readonly user?: components["schemas"]["User"];
            user_id?: number;
            /** Format: decimal */
            amount?: string;
            /** Format: date */
            month?: string | null;
            /** Format: date-time */
            readonly created_at?: string;
        };
// ...
        Role: {
            readonly id: number;
            code: string;
            name: string;
        };
// ...
        UserSalary: {
            readonly id: number;
            readonly user: components["schemas"]["User"];
            user_id: number;
            /** Format: decimal */
            amount: string;
            /** Format: date */
            month?: string | null;
            /** Format: date-time */
            readonly created_at: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}